+++ Foodle/Foodle/Cell/LoginViewCell.swift
@@ -10,16 +10,15 @@ import UIKit
 class LoginViewCell: UIViewController {
     @IBOutlet var loginLabel: UILabel!
     @IBOutlet var loginLabel2: UILabel!
-    @IBOutlet var kakaoImg: UIImageView!
-    @IBOutlet var naverImg: UIImageView!
-    @IBOutlet var appleImg: UIImageView!
+    @IBOutlet var kakaoButton: UIButton!
+    @IBOutlet var naverButton: UIButton!
+    @IBOutlet var appleButton: UIButton!
     @IBOutlet var privacyLabel: UILabel!
+    
+    
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        kakaoImg.image = UIImage(named: "icon_kakao.png")
-        naverImg.image = UIImage(named: "icon_naver.png")
-        appleImg.image = UIImage(named: "icon_apple.png")
         
         // PrivacyLabel에 밑줄 추가
         let attributedString = NSMutableAttributedString(string: "개인정보 이용 처리 방침")
